requirements.txt
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.5.0
astroid==2.9.3
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
cached-property==1.5.2
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.0
defusedxml==0.7.1
diff-match-patch==20200713
discord-webhook==0.14.0
Django==3.2.12
django-analytical==3.0.0
django-braces==1.14.0
django-colorfield==0.4.3
django-compressor==2.4
django-cors-headers==3.7.0
django-debug-toolbar==3.4.0
django-dbbackup==3.3.0
django-extensions==3.1.5
django-filter==2.4.0
django-imagekit==4.0.2
django-jalali==4.1.2
django-mathfilters==1.0.0
django-modeladmin-reorder==0.3.1
django-oauth-toolkit==1.4.0
django-prometheus==2.1.0
django-rename-app==0.1.3
django-simple-history==3.0.0
django-tinymce==2.8.0
django-import-export==2.8.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
docutils==0.17.1
drf-yasg==1.20.0
et-xmlfile==1.1.0
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.0.4
idna==2.8
imagesize==1.3.0
importlib-metadata==4.11.1
inflection==0.5.1
ipython==7.27.0
isodate==0.6.0
isort==5.10.1
itypes==1.2.0
jdatetime==4.1.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==3.0.3
kavenegar==1.1.2
lazy-object-proxy==1.7.1
lxml==4.6.3
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.2
oauthlib==3.1.1
odfpy==1.4.1
openapi-codec==1.3.2
packaging==21.3
pandas==1.3.3
parso==0.8.2
pathspec==0.9.0
pbr==5.8.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.0.1
platformdirs==2.5.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
PyJWT==1.7.0
pylint==2.12.2
pylint-django==2.5.2
pylint-plugin-utils==0.7
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==6.0
rcssmin==1.0.6
requests==2.28.0
requests-toolbelt==0.9.1
rjsmin==1.1.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sentry-sdk==1.5.12
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==4.4.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
stevedore==3.5.0
tablib==3.2.1
toml==0.10.2
tomli==2.0.1
traitlets==5.1.0
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.25.11
wcwidth==0.2.5
wrapt==1.13.3
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
zeep==3.4.0
zipp==3.7.0
